+++ internal/batches/service/service.go
@@ -10,6 +10,7 @@ import (
 	"path"
 	"regexp"
 	"strings"
+	"sync"
 
 	"github.com/cockroachdb/errors"
 
@@ -29,6 +30,7 @@ type Service struct {
 	client           api.Client
 	features         batches.FeatureFlags
 	imageCache       *docker.ImageCache
+	parallelism      int
 }
 
 type Opts struct {
@@ -36,6 +38,7 @@ type Opts struct {
 	AllowIgnored     bool
 	AllowFiles       bool
 	Client           api.Client
+	Parallelism      int
 }
 
 var (
@@ -43,12 +46,18 @@ var (
 )
 
 func New(opts *Opts) *Service {
+	parallelism := opts.Parallelism
+	if parallelism < 1 {
+		parallelism = 1
+	}
+
 	return &Service{
 		allowUnsupported: opts.AllowUnsupported,
 		allowIgnored:     opts.AllowIgnored,
 		allowFiles:       opts.AllowFiles,
 		client:           opts.Client,
 		imageCache:       docker.NewImageCache(),
+		parallelism:      parallelism,
 	}
 }
 
@@ -157,20 +166,74 @@ func (svc *Service) CreateChangesetSpec(ctx context.Context, spec *batcheslib.Ch
 // Progress information is reported back to the given progress function: perc
 // will be a value between 0.0 and 1.0, inclusive.
 func (svc *Service) EnsureDockerImages(ctx context.Context, spec *batcheslib.BatchSpec, progress func(done, total int)) (map[string]docker.Image, error) {
-	total := len(spec.Steps)
+	// Figure out the image names used in the batch spec.
+	names := map[string]struct{}{}
+	for i := range spec.Steps {
+		names[spec.Steps[i].Container] = struct{}{}
+	}
+
+	// Set up our progress reporting.
+	total := len(names)
 	progress(0, total)
 
-	// TODO: this _really_ should be parallelised, since the image cache takes
-	// care to only pull the same image once.
-	images := make(map[string]docker.Image)
-	for i := range spec.Steps {
-		img, err := svc.EnsureImage(ctx, spec.Steps[i].Container)
-		if err != nil {
-			return nil, err
-		}
-		images[spec.Steps[i].Container] = img
+	// Set up the channels that will be used in the parallel goroutines handling
+	// the pulls.
+	type image struct {
+		name  string
+		image docker.Image
+		err   error
+	}
+	complete := make(chan image)
+	inputs := make(chan string, total)
 
-		progress(i+1, total)
+	// Set up a worker context that we can use to terminate the workers if an
+	// error occurs.
+	workerCtx, cancel := context.WithCancel(ctx)
+	defer cancel()
+
+	// Spawn worker goroutines to call EnsureImage on each image name.
+	var wg sync.WaitGroup
+	for i := 0; i < svc.parallelism; i++ {
+		wg.Add(1)
+		go func() {
+			for name := range inputs {
+				img, err := svc.EnsureImage(workerCtx, name)
+				complete <- image{
+					name:  name,
+					image: img,
+					err:   err,
+				}
+			}
+			wg.Done()
+		}()
+	}
+
+	go func() {
+		wg.Wait()
+		close(complete)
+	}()
+
+	// Send the image names to the worker goroutines.
+	go func() {
+		for name := range names {
+			inputs <- name
+		}
+		close(inputs)
+	}()
+
+	// Receive the results of the image pulls and build the return value.
+	i := 0
+	images := make(map[string]docker.Image)
+	for image := range complete {
+		if image.err != nil {
+			// If EnsureImage errored, then we'll early return here and let the
+			// worker context clean things up.
+			return nil, image.err
+		}
+
+		images[image.name] = image.image
+		i += 1
+		progress(i, total)
 	}
 
 	return images, nil